Automatyczny Stock Market 1.0/Algorytm Uczenia Kwantowego LLM/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
bcrypt
passlib[argon2]
pydantic-settings
python-multipart
jinja2
python-dotenv
google-generativeai
ta
orjson
alembic
slowapi
PyJWT
prometheus-client
pyotp
qrcode
httpx
Pillow
sentry-sdk[fastapi]
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-system-metrics
redis
pywebpush
asyncpg
email-validator

# New dependencies for enhanced functionality
aiohttp>=3.9.0
aiofiles>=23.2.1
cryptography>=41.0.0
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
textblob>=0.17.1
vaderSentiment>=3.3.2
python-dateutil>=2.8.2
schedule>=1.2.0
websockets>=12.0
python-telegram-bot>=20.7
tweepy>=4.14.0
praw>=7.7.1
feedparser>=6.0.10
newspaper3k>=0.2.8
yfinance>=0.2.24
ccxt>=4.2.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.17.0
dash>=2.14.0
streamlit>=1.28.0

supabase
anthropic
